Create README.md
The project is an end to end deep learning project in agriculture domain. 
Farmers every year face economic loss and crop waste due to various diseases in Potato plant
There are two type to diseases in this
1 Early blight: caused by the fungus
2 Late blight: casued by spefic microrganisum
if farfer can detect this elarly then he will pervent the economic loss.
We aregoing  build a Convolutional Neural Network Model which will take input as images 
It will clarify and test the image and give the output regarding the disease associated with the image.