Update README.md
The project is an end to end deep learning project in the agriculture domain. 
Farmers every year face economic loss and crop waste because of various diseases in potato plant
There are two type to diseases in this
1 Early blight: caused by the fungus
2 Late blight: caused by specific microorganisms
if farmer can detect this elarly then they will prevent economic loss.
We are going build a Convolutional Neural Network Model which will take input as images 
It will clarify and test the image and give the output regarding the disease associated with the image.